# packages/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cable Platform API",
    "version": "1.0.0",
    "description": "API for cable and harness design pipeline"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/synthesis/propose": {
      "post": {
        "summary": "Generate synthesis proposal from draft",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "draft_id"
                ],
                "properties": {
                  "draft_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Synthesis proposal generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SynthesisProposal"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Draft not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/synthesis/accept": {
      "post": {
        "summary": "Accept synthesis proposal and create assembly",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "proposal_id"
                ],
                "properties": {
                  "proposal_id": {
                    "type": "string"
                  },
                  "locks": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assembly created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "assembly_id",
                    "schema_hash"
                  ],
                  "properties": {
                    "assembly_id": {
                      "type": "string"
                    },
                    "schema_hash": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Proposal not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/synthesis/recompute": {
      "post": {
        "summary": "Recompute synthesis proposal with locks",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "draft_id"
                ],
                "properties": {
                  "draft_id": {
                    "type": "string"
                  },
                  "locks": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Synthesis proposal recomputed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SynthesisProposal"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Draft not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PartRef": {
        "type": "object",
        "required": [
          "mpn"
        ],
        "properties": {
          "mpn": {
            "type": "string"
          },
          "family": {
            "type": "string"
          },
          "series": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "WirelistRow": {
        "type": "object",
        "required": [
          "circuit",
          "conductor"
        ],
        "properties": {
          "circuit": {
            "type": "string"
          },
          "conductor": {
            "type": "integer",
            "minimum": 1
          },
          "endA_pin": {
            "type": "string"
          },
          "endB_pin": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "shield": {
            "type": "string",
            "enum": [
              "none",
              "fold_back",
              "isolated",
              "pigtail"
            ]
          }
        }
      },
      "BomLine": {
        "type": "object",
        "required": [
          "ref",
          "qty",
          "role"
        ],
        "properties": {
          "ref": {
            "$ref": "#/components/schemas/PartRef"
          },
          "qty": {
            "type": "integer",
            "minimum": 1
          },
          "role": {
            "type": "string",
            "enum": [
              "primary",
              "alternate"
            ]
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "SynthesisProposal": {
        "type": "object",
        "required": [
          "proposal_id",
          "draft_id",
          "cable",
          "conductors",
          "endpoints",
          "shield",
          "wirelist",
          "bom",
          "warnings",
          "errors",
          "explain"
        ],
        "properties": {
          "proposal_id": {
            "type": "string"
          },
          "draft_id": {
            "type": "string"
          },
          "cable": {
            "type": "object",
            "required": [
              "primary",
              "alternates"
            ],
            "properties": {
              "primary": {
                "$ref": "#/components/schemas/PartRef"
              },
              "alternates": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PartRef"
                }
              }
            }
          },
          "conductors": {
            "type": "object",
            "properties": {
              "count": {
                "type": "integer",
                "minimum": 1
              },
              "awg": {
                "type": "integer"
              },
              "color_map": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "ribbon": {
                "type": "object",
                "required": [
                  "pitch_in",
                  "ways",
                  "red_stripe"
                ],
                "properties": {
                  "pitch_in": {
                    "type": "number",
                    "enum": [
                      0.025,
                      0.05
                    ]
                  },
                  "ways": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "red_stripe": {
                    "type": "boolean"
                  }
                }
              }
            }
          },
          "endpoints": {
            "type": "object",
            "required": [
              "endA",
              "endB"
            ],
            "properties": {
              "endA": {
                "type": "object",
                "required": [
                  "connector",
                  "termination"
                ],
                "properties": {
                  "connector": {
                    "$ref": "#/components/schemas/PartRef"
                  },
                  "termination": {
                    "type": "string",
                    "enum": [
                      "crimp",
                      "idc",
                      "ring_lug",
                      "solder"
                    ]
                  },
                  "contacts": {
                    "type": "object",
                    "required": [
                      "primary",
                      "alternates"
                    ],
                    "properties": {
                      "primary": {
                        "$ref": "#/components/schemas/PartRef"
                      },
                      "alternates": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/PartRef"
                        }
                      }
                    }
                  },
                  "accessories": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PartRef"
                    }
                  }
                }
              },
              "endB": {
                "type": "object",
                "required": [
                  "connector",
                  "termination"
                ],
                "properties": {
                  "connector": {
                    "$ref": "#/components/schemas/PartRef"
                  },
                  "termination": {
                    "type": "string",
                    "enum": [
                      "crimp",
                      "idc",
                      "ring_lug",
                      "solder"
                    ]
                  },
                  "contacts": {
                    "type": "object",
                    "required": [
                      "primary",
                      "alternates"
                    ],
                    "properties": {
                      "primary": {
                        "$ref": "#/components/schemas/PartRef"
                      },
                      "alternates": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/PartRef"
                        }
                      }
                    }
                  },
                  "accessories": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PartRef"
                    }
                  }
                }
              }
            }
          },
          "shield": {
            "type": "object",
            "required": [
              "type",
              "drain_policy"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "none",
                  "foil",
                  "braid",
                  "foil_braid"
                ]
              },
              "drain_policy": {
                "type": "string",
                "enum": [
                  "isolated",
                  "fold_back",
                  "pigtail"
                ]
              }
            }
          },
          "wirelist": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WirelistRow"
            }
          },
          "bom": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BomLine"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "explain": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}